% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyEpidemic.R
\name{calcValue}
\alias{calcValue}
\title{calculate phenotypic values}
\usage{
calcValue(env, gene, e, GEValues)
}
\arguments{
\item{env}{integer vector containing environmental types for each individual}

\item{gene}{integer vectory containing genotype indices for each individual}

\item{e}{numeric vector containing the special environmental effects for each
individual}

\item{GEValues}{matrix containing the genotype by environment expected
phenotypic values}
}
\value{
vector of phenotypic values
}
\description{
calculate phenotypic values
}

