% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyEpidemic.R
\name{decomposeTrait}
\alias{decomposeTrait}
\title{Decompose trait values according to the model z=G+I+E+epsilon}
\usage{
decomposeTrait(data, GEValues, copy = TRUE)
}
\arguments{
\item{data}{a data.table returned by extractPop}

\item{GEValues}{a matrix of GEValues}

\item{copy}{logical indicating whether a copy of the original data.table
should be returned (TRUE by default)}
}
\description{
Decompose trait values according to the model z=G+I+E+epsilon
}

