% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyEpidemic.R
\name{generateGenotypes}
\alias{generateGenotypes}
\title{Possible genotypes based on numbers of alleles at QTLs.}
\usage{
generateGenotypes(numsAllelesAtSites)
}
\arguments{
\item{numsAllelesAtSites}{integer vector specifying the number of alleles at each site,
the length of the vector defining the number of QTLs.}
}
\description{
Possible genotypes based on numbers of alleles at QTLs.
}

