% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyEpidemic.R
\name{generateRateTransTemplate}
\alias{generateRateTransTemplate}
\title{Generate a state transition template matrix based on numbers of alleles at
QTLs.}
\usage{
generateRateTransTemplate(numsAllelesAtSites, nameRows = FALSE)
}
\arguments{
\item{numsAllelesAtSites}{integer vector specifying the number of alleles at
each site, the length of the vector defining the number of QTLs.}

\item{nameRows}{logical indicating whether the returned matrix should have
named rows (for explicative purpose only, default FALSE). The column names
are different for each row, i.e. the row-names excluding the genotype
corresponding to that specific row.}
}
\description{
Generate a state transition template matrix based on numbers of alleles at
QTLs.
}
\details{
each element of the matrix is a multiplier for the corresponding
  genotype transition rate (see function simulateEpidemic in the patherit
  package). This multiplier is nonzero only for transitions that need exactly
  one mutation (haming distance = 1) in order to convert one genotype into
  another. The
}

