% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyEpidemic.R
\name{newgennxNorm}
\alias{newgennxNorm}
\title{simulating evolution on continuous time scale
sde : numeric or vector of standard deviations of the environmental effects for each allele
N : number of individuals
nGtps : number of alleles
sde : a numeric or a numeric matrix, sde[env, g]: standard dev of the special env. effect for <env, g>}
\usage{
newgennxNorm(pe, sde, pg.init, N, eUniqForEachG = FALSE)
}
\description{
simulating evolution on continuous time scale
sde : numeric or vector of standard deviations of the environmental effects for each allele
N : number of individuals
nGtps : number of alleles
sde : a numeric or a numeric matrix, sde[env, g]: standard dev of the special env. effect for <env, g>
}

